Move path combinators into separate module
